modificare functie post review pt a afisa si reviews pe care le primeste pt proprietatile pe care le are si cele pe care la posteaza el

+++ src/main/java/com/example/backend_springboot/dtos/reviewDTO/GetReviewForUserDTO.java
@@ -1,5 +1,6 @@
 package com.example.backend_springboot.dtos.reviewDTO;
 
+import jnr.ffi.annotations.In;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,5 +15,5 @@ public class GetReviewForUserDTO {
     private UUID idReview;
     private LocalDate date;
     private String comment;
-    private Double rating;
+    private Integer rating;
 }
